third_party/blink/web_tests/platform/generic/http/tests/websocket/send-object-tostring-check-expected.txt

WebSocket: Object's toString method should be called only once.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS event.data is "Goodbye"
PASS testObj.callCounter is 1
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
